pyproject.toml
[tool.poetry]
name = "winston"
version = "0.1.0"
description = "Utilities & Moderation Tools for Discord"
authors = ["MrArkon <mrarkon@outlook.com>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = {extras = ["speed"], version = "^2.1.0"}
tomli = "^2.0.1"
jishaku = "^2.5.1"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
combine_star = true
star_first = true
include_trailing_comma = true
dedup_headings = true
known_first_party = "poetry"

[tool.pyright]
reportOptionalMemberAccess = false
reportShadowedImports = false
